Updated Poll Logic; Updated Heisman Logic

+++ src/main/java/Simulation/Team.java
@@ -782,11 +782,11 @@ public class Team {
         teamOffTalent = getOffTalent();
         teamDefTalent = getDefTalent();
 
-        int preseasonBias = 12 - (wins + losses);
+        double preseasonBias = 12 - (wins + losses);
         if (preseasonBias < 0) preseasonBias = 0;
         preseasonBias = preseasonBias/15;
         teamPollScore =
-                preseasonBias * (teamOffTalent + teamDefTalent + (teamPrestige/2) + (confPrestige/3)) +
+                (int)(preseasonBias * (teamOffTalent + teamDefTalent + (teamPrestige/2) + (confPrestige/3))) +
                         (offRating + defRating + teamStrengthOfWins - teamStrengthOfLosses +
                         1000)/2;
